requirements.txt
django==5.0.3
websocket-client==1.7.0
redis==5.0.3
flake8==7.0.0
djangorestframework==3.15.1
django-rest-swagger==2.2.0
drf-yasg==1.21.7
pytest==8.0.2
pytest-django==4.8.0
